封装确认框组件并应用

+++ erabbit-pc-vue-project/src/views/cart/index.vue
@@ -9,7 +9,12 @@
         <table>
           <thead>
             <tr>
-              <th width="120"><XtxCheckbox @change="checkAll" :modelValue="$store.getters['cart/isCheckAll']">全选</XtxCheckbox></th>
+              <th width="120">
+                <XtxCheckbox
+                  @change="checkAll"
+                  :modelValue="$store.getters['cart/isCheckAll']"
+                >全选</XtxCheckbox>
+              </th>
               <th width="400">商品信息</th>
               <th width="220">单价</th>
               <th width="180">数量</th>
@@ -24,11 +29,22 @@
                 <CartNone></CartNone>
               </td>
             </tr>
-            <tr v-for="goods in $store.getters['cart/validList']" :key="goods.skuId">
-              <td><XtxCheckbox @change="($event)=>checkOne(goods.skuId, $event)" :modelValue="goods.selected" /></td>
+            <tr
+              v-for="goods in $store.getters['cart/validList']"
+              :key="goods.skuId"
+            >
+              <td>
+                <XtxCheckbox
+                  @change="($event)=>checkOne(goods.skuId, $event)"
+                  :modelValue="goods.selected"
+                />
+              </td>
               <td>
                 <div class="goods">
-                  <RouterLink :to="`/product/${goods.id}`"><img :src="goods.picture" alt=""></RouterLink>
+                  <RouterLink :to="`/product/${goods.id}`"><img
+                      :src="goods.picture"
+                      alt=""
+                    ></RouterLink>
                   <div>
                     <p class="name ellipsis">{{goods.name}}</p>
                     <!-- 选择规格组件 -->
@@ -40,37 +56,61 @@
                 <p v-if="goods.price - goods.nowPrice > 0">比加入时降价 <span class="red">&yen;{{goods.price - goods.nowPrice}}</span></p>
               </td>
               <td class="tc">
-                <XtxNumbox :modelValue="goods.count"/>
+                <XtxNumbox :modelValue="goods.count" />
               </td>
               <td class="tc">
                 <p class="f16 red">&yen;{{Math.round(goods.nowPrice * 100) * goods.count / 100}}</p>
               </td>
               <td class="tc">
                 <p><a href="javascript:;">移入收藏夹</a></p>
-                <p><a class="green" href="javascript:;" @click="deleteCart(goods.skuId)">删除</a></p>
+                <p><a
+                    class="green"
+                    href="javascript:;"
+                    @click="deleteCart(goods.skuId)"
+                  >删除</a></p>
                 <p><a href="javascript:;">找相似</a></p>
               </td>
             </tr>
           </tbody>
           <!-- 无效商品 -->
           <tbody v-if="$store.getters['cart/invalidList'].length">
-            <tr><td colspan="6"><h3 class="tit">失效商品</h3></td></tr>
-            <tr v-for="goods in $store.getters['cart/invalidList']" :key="goods.skuId">
-              <td><XtxCheckbox style="color:#eee;" /></td>
+            <tr>
+              <td colspan="6">
+                <h3 class="tit">失效商品</h3>
+              </td>
+            </tr>
+            <tr
+              v-for="goods in $store.getters['cart/invalidList']"
+              :key="goods.skuId"
+            >
+              <td>
+                <XtxCheckbox style="color:#eee;" />
+              </td>
               <td>
                 <div class="goods">
-                  <RouterLink to="/"><img :src="goods.picture" alt=""></RouterLink>
+                  <RouterLink to="/"><img
+                      :src="goods.picture"
+                      alt=""
+                    ></RouterLink>
                   <div>
                     <p class="name ellipsis">{{goods.name}}</p>
                     <p class="attr">{{goods.attrsText}}</p>
                   </div>
                 </div>
               </td>
-              <td class="tc"><p>&yen;{{goods.nowPrice}}</p></td>
-              <td class="tc">{{goods.count}}</td>
-              <td class="tc"><p>&yen;{{Math.round(goods.nowPrice * 100) * goods.count / 100}}</p></td>
               <td class="tc">
-                <p><a class="green" href="javascript:;" @click="deleteCart(goods.skuId)">删除</a></p>
+                <p>&yen;{{goods.nowPrice}}</p>
+              </td>
+              <td class="tc">{{goods.count}}</td>
+              <td class="tc">
+                <p>&yen;{{Math.round(goods.nowPrice * 100) * goods.count / 100}}</p>
+              </td>
+              <td class="tc">
+                <p><a
+                    class="green"
+                    href="javascript:;"
+                    @click="deleteCart(goods.skuId)"
+                  >删除</a></p>
                 <p><a href="javascript:;">找相似</a></p>
               </td>
             </tr>
@@ -80,7 +120,10 @@
       <!-- 操作栏 -->
       <div class="action">
         <div class="batch">
-          <XtxCheckbox :modelValue="$store.getters['cart/isCheckAll']" @change="checkAll">全选</XtxCheckbox>
+          <XtxCheckbox
+            :modelValue="$store.getters['cart/isCheckAll']"
+            @change="checkAll"
+          >全选</XtxCheckbox>
           <a href="javascript:;">删除商品</a>
           <a href="javascript:;">移入收藏夹</a>
           <a href="javascript:;">清空失效商品</a>
@@ -97,32 +140,40 @@
   </div>
 </template>
 <script>
-import GoodRelevant from '@/views/goods/components/goods-relevant'
-import { useStore } from "vuex"
+import GoodRelevant from "@/views/goods/components/goods-relevant";
+import { useStore } from "vuex";
 import Message from "@/components/library/Message";
-import CartNone from "./components/cart-none.vue"
+import CartNone from "./components/cart-none.vue";
+import Confirm from "@/components/library/Confirm";
 export default {
-  name: 'XtxCartPage',
+  name: "XtxCartPage",
   components: { GoodRelevant, CartNone },
   setup () {
     const store = useStore();
     // 单选
     const checkOne = (skuId, selected) => {
-      store.dispatch('cart/updateCart', { skuId, selected })
-    }
+      store.dispatch("cart/updateCart", { skuId, selected });
+    };
     // 全选
     const checkAll = (selected) => {
-      store.dispatch('cart/checkAllCart', selected)
-    }
+      store.dispatch("cart/checkAllCart", selected);
+    };
     // 删除商品
     const deleteCart = (skuId) => {
-      store.dispatch('cart/deleteCart', skuId).then(() => {
-        Message({ type: "success", text: "删除商品成功" });
-      });
-    }
-    return { checkOne, checkAll, deleteCart }
-  }
-}
+      Confirm({ text: "亲，您是否确认删除商品" })
+        .then(() => {
+          // console.log("确认");
+          store.dispatch("cart/deleteCart", skuId).then(() => {
+            Message({ type: "success", text: "删除商品成功" });
+          });
+        })
+        .catch((e) => {
+          // console.log("取消");
+        });
+    };
+    return { checkOne, checkAll, deleteCart };
+  },
+};
 </script>
 <style scoped lang="less">
 .tc {
@@ -136,7 +187,7 @@ export default {
   color: @priceColor;
 }
 .green {
-  color: @xtxColor
+  color: @xtxColor;
 }
 .f16 {
   font-size: 16px;
@@ -195,7 +246,8 @@ export default {
       border-spacing: 0;
       border-collapse: collapse;
       line-height: 24px;
-      th,td{
+      th,
+      td {
         padding: 10px;
         border-bottom: 1px solid #f5f5f5;
         &:first-child {
